+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Acknowledge the new json format for events in the Assistant [#5255](https://github.com/rokwire/illinois-app/issues/5255).
 
 ## [7.2.3] - 2025-06-26
 ### Added